Moved units (de)serialization to a separate file.
