+++ src/shammodels/sph/include/shammodels/sph/SolverConfig.hpp
@@ -28,6 +28,7 @@
 #include "shammodels/common/EOSConfig.hpp"
 #include "shammodels/common/ExtForceConfig.hpp"
 #include "shammodels/sph/config/MHDConfig.hpp"
+#include "shamrock/io/units_json.hpp"
 #include "shamrock/patch/PatchDataLayout.hpp"
 #include "shamsys/NodeInstance.hpp"
 #include "shamsys/legacy/log.hpp"
@@ -686,46 +687,6 @@ struct shammodels::sph::SolverConfig {
     void set_ghost_layout(shamrock::patch::PatchDataLayout &ghost_layout);
 };
 
-namespace shamunits {
-
-    /**
-     * @brief Converts a UnitSystem object to a JSON object.
-     *
-     * @param j The JSON object to be populated.
-     * @param p The UnitSystem object to be converted.
-     */
-    template<class Tscal>
-    inline void to_json(nlohmann::json &j, const UnitSystem<Tscal> &p) {
-        j = nlohmann::json{
-            {"unit_time", p.s_inv},
-            {"unit_length", p.m_inv},
-            {"unit_mass", p.kg_inv},
-            {"unit_current", p.A_inv},
-            {"unit_temperature", p.K_inv},
-            {"unit_qte", p.mol_inv},
-            {"unit_lumint", p.cd_inv}};
-    }
-
-    /**
-     * @brief Deserializes a UnitSystem object from a JSON object.
-     *
-     * @param j The JSON object to deserialize from.
-     * @param p The UnitSystem object to populate.
-     */
-    template<class Tscal>
-    inline void from_json(const nlohmann::json &j, UnitSystem<Tscal> &p) {
-        p = UnitSystem<Tscal>(
-            j.at("unit_time").get<Tscal>(),
-            j.at("unit_length").get<Tscal>(),
-            j.at("unit_mass").get<Tscal>(),
-            j.at("unit_current").get<Tscal>(),
-            j.at("unit_temperature").get<Tscal>(),
-            j.at("unit_qte").get<Tscal>(),
-            j.at("unit_lumint").get<Tscal>());
-    }
-
-} // namespace shamunits
-
 namespace shammodels::sph {
 
     /**
@@ -753,44 +714,6 @@ namespace shammodels::sph {
         j.at("cfl_cour").get_to<Tscal>(p.cfl_cour);
         j.at("cfl_force").get_to<Tscal>(p.cfl_force);
         j.at("cfl_multiplier_stiffness").get_to<Tscal>(p.cfl_multiplier_stiffness);
-    }
-
-    /**
-     * @brief Converts an optional value to a JSON object.
-     *
-     * @param j The JSON object to be populated.
-     * @param p The optional value to be converted.
-     *
-     * @return None
-     *
-     * @throws std::bad_optional_access if p is not engaged
-     */
-    template<class T>
-    inline void to_json_optional(nlohmann::json &j, const std::optional<T> &p) {
-        if (p) {
-            j = *p;
-        } else {
-            j = {};
-        }
-    }
-
-    /**
-     * @brief Deserializes an optional value from a JSON object.
-     *
-     * @param j The JSON object to deserialize from.
-     * @param p The optional value to populate.
-     *
-     * @return None
-     *
-     * @throws std::bad_optional_access if j is not a valid JSON object
-     */
-    template<class T>
-    inline void from_json_optional(const nlohmann::json &j, std::optional<T> &p) {
-        if (j.is_null()) {
-            p = std::nullopt;
-        } else {
-            p = j.get<T>();
-        }
     }
 
     /**
